# ⚡ FEN parsing optimization: en-passant parsing