Optimize `ParseEnpassant`, which happens to represent the.. 66.5% of ParseFEN!

Done by replacing enum `Enum.TryParse` with ASCII char-based parsing.

Flame graph before:
<img width="1510" height="167" alt="image" src="https://github.com/user-attachments/assets/43558641-fd35-474a-ad49-65c340e755ba" />

Flame graph after:
<img width="1641" height="215" alt="image" src="https://github.com/user-attachments/assets/74d0a001-ccf3-4a78-8648-66d3f9c21e01" />

<details>

Increase of allocations is caused due to of the `ParseFENResult` struct introduced as part of FRC support, and not this PR.

<Summary>Benchmarks:</Summary>

```
BenchmarkDotNet v0.14.0, Ubuntu 24.04.3 LTS (Noble Numbat)
AMD EPYC 7763, 1 CPU, 4 logical and 2 physical cores
.NET SDK 9.0.305
  [Host]     : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2

| Method                                | fen                  | Mean       | Error    | StdDev   | Ratio | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------------- |--------------------- |-----------:|---------:|---------:|------:|-------:|-------:|----------:|------------:|
| ParseFEN_Original                     | 8/k7/(...)- 0 1 [39] | 2,163.7 ns | 19.92 ns | 18.63 ns |  1.00 | 0.1755 |      - |    2961 B |        1.00 |
| ParseFEN_Improved1                    | 8/k7/(...)- 0 1 [39] | 1,955.1 ns | 20.14 ns | 18.84 ns |  0.90 | 0.1602 |      - |    2705 B |        0.91 |
| ParseFEN_Base2                        | 8/k7/(...)- 0 1 [39] | 1,993.1 ns | 23.42 ns | 21.91 ns |  0.92 | 0.1640 |      - |    2761 B |        0.93 |
| ParseFEN_NoRegex                      | 8/k7/(...)- 0 1 [39] |   885.5 ns |  8.15 ns |  7.63 ns |  0.41 | 0.0286 |      - |     480 B |        0.16 |
| ParseFEN_OptimizedParseBoard          | 8/k7/(...)- 0 1 [39] |   309.3 ns |  1.96 ns |  1.74 ns |  0.14 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedPopulateOccupancies | 8/k7/(...)- 0 1 [39] |   328.5 ns |  0.65 ns |  0.58 ns |  0.15 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedParseEnPassant      | 8/k7/(...)- 0 1 [39] |   272.4 ns |  1.56 ns |  1.46 ns |  0.13 | 0.0348 | 0.0010 |     584 B |        0.20 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r2q1r(...)- 0 9 [68] | 2,612.0 ns | 21.50 ns | 20.11 ns |  1.00 | 0.1869 |      - |    3161 B |        1.00 |
| ParseFEN_Improved1                    | r2q1r(...)- 0 9 [68] | 2,434.7 ns |  9.81 ns |  9.17 ns |  0.93 | 0.1717 |      - |    2905 B |        0.92 |
| ParseFEN_Base2                        | r2q1r(...)- 0 9 [68] | 2,504.4 ns |  6.07 ns |  5.38 ns |  0.96 | 0.1793 |      - |    3017 B |        0.95 |
| ParseFEN_NoRegex                      | r2q1r(...)- 0 9 [68] | 1,255.2 ns |  1.94 ns |  1.72 ns |  0.48 | 0.0362 |      - |     624 B |        0.20 |
| ParseFEN_OptimizedParseBoard          | r2q1r(...)- 0 9 [68] |   393.0 ns |  0.59 ns |  0.53 ns |  0.15 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r2q1r(...)- 0 9 [68] |   413.8 ns |  0.41 ns |  0.32 ns |  0.16 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r2q1r(...)- 0 9 [68] |   333.7 ns |  0.62 ns |  0.52 ns |  0.13 | 0.0348 | 0.0010 |     584 B |        0.18 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r3k2r(...)- 0 1 [68] | 2,477.0 ns |  8.02 ns |  7.11 ns |  1.00 | 0.1831 |      - |    3081 B |        1.00 |
| ParseFEN_Improved1                    | r3k2r(...)- 0 1 [68] | 2,587.7 ns |  5.15 ns |  4.57 ns |  1.04 | 0.1678 |      - |    2817 B |        0.91 |
| ParseFEN_Base2                        | r3k2r(...)- 0 1 [68] | 2,384.0 ns | 25.45 ns | 21.25 ns |  0.96 | 0.1717 |      - |    2929 B |        0.95 |
| ParseFEN_NoRegex                      | r3k2r(...)- 0 1 [68] | 1,125.9 ns | 11.10 ns | 10.38 ns |  0.45 | 0.0324 |      - |     552 B |        0.18 |
| ParseFEN_OptimizedParseBoard          | r3k2r(...)- 0 1 [68] |   394.8 ns |  2.32 ns |  2.17 ns |  0.16 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r3k2r(...)- 0 1 [68] |   419.6 ns |  2.05 ns |  1.82 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r3k2r(...)- 0 1 [68] |   326.0 ns |  1.72 ns |  1.61 ns |  0.13 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r3k2r(...)- 0 1 [68] | 2,442.3 ns |  5.34 ns |  4.73 ns |  1.00 | 0.1831 |      - |    3081 B |        1.00 |
| ParseFEN_Improved1                    | r3k2r(...)- 0 1 [68] | 2,441.4 ns | 20.05 ns | 18.75 ns |  1.00 | 0.1678 |      - |    2817 B |        0.91 |
| ParseFEN_Base2                        | r3k2r(...)- 0 1 [68] | 2,395.8 ns | 16.90 ns | 14.98 ns |  0.98 | 0.1717 |      - |    2929 B |        0.95 |
| ParseFEN_NoRegex                      | r3k2r(...)- 0 1 [68] | 1,154.2 ns |  2.26 ns |  2.00 ns |  0.47 | 0.0324 |      - |     552 B |        0.18 |
| ParseFEN_OptimizedParseBoard          | r3k2r(...)- 0 1 [68] |   395.0 ns |  1.84 ns |  1.72 ns |  0.16 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r3k2r(...)- 0 1 [68] |   410.1 ns |  0.53 ns |  0.45 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r3k2r(...)- 0 1 [68] |   345.8 ns |  1.32 ns |  1.17 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rnbqk(...)6 0 1 [67] | 2,436.6 ns |  2.64 ns |  2.34 ns |  1.00 | 0.1831 |      - |    3073 B |        1.00 |
| ParseFEN_Improved1                    | rnbqk(...)6 0 1 [67] | 2,289.6 ns |  9.62 ns |  8.04 ns |  0.94 | 0.1640 |      - |    2801 B |        0.91 |
| ParseFEN_Base2                        | rnbqk(...)6 0 1 [67] | 2,329.7 ns | 20.01 ns | 18.71 ns |  0.96 | 0.1717 |      - |    2905 B |        0.95 |
| ParseFEN_NoRegex                      | rnbqk(...)6 0 1 [67] | 1,105.0 ns |  5.54 ns |  5.18 ns |  0.45 | 0.0305 |      - |     528 B |        0.17 |
| ParseFEN_OptimizedParseBoard          | rnbqk(...)6 0 1 [67] |   412.3 ns |  0.99 ns |  0.87 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | rnbqk(...)6 0 1 [67] |   388.8 ns |  2.03 ns |  1.90 ns |  0.16 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | rnbqk(...)6 0 1 [67] |   335.5 ns |  2.57 ns |  2.28 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rnbqk(...)- 0 1 [56] | 1,802.3 ns |  3.97 ns |  3.32 ns |  1.00 | 0.1640 | 0.0019 |    2761 B |        1.00 |
| ParseFEN_Improved1                    | rnbqk(...)- 0 1 [56] | 1,709.8 ns |  3.78 ns |  3.35 ns |  0.95 | 0.1488 | 0.0019 |    2496 B |        0.90 |
| ParseFEN_Base2                        | rnbqk(...)- 0 1 [56] | 1,798.2 ns |  9.45 ns |  8.84 ns |  1.00 | 0.1545 | 0.0019 |    2584 B |        0.94 |
| ParseFEN_NoRegex                      | rnbqk(...)- 0 1 [56] |   634.8 ns |  1.00 ns |  0.83 ns |  0.35 | 0.0153 |      - |     264 B |        0.10 |
| ParseFEN_OptimizedParseBoard          | rnbqk(...)- 0 1 [56] |   386.9 ns |  0.53 ns |  0.49 ns |  0.21 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedPopulateOccupancies | rnbqk(...)- 0 1 [56] |   357.9 ns |  0.89 ns |  0.74 ns |  0.20 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedParseEnPassant      | rnbqk(...)- 0 1 [56] |   340.7 ns |  0.59 ns |  0.49 ns |  0.19 | 0.0348 | 0.0010 |     584 B |        0.21 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rq2k2(...)- 0 1 [71] | 2,746.8 ns |  3.55 ns |  3.15 ns |  1.00 | 0.1869 |      - |    3169 B |        1.00 |
| ParseFEN_Improved1                    | rq2k2(...)- 0 1 [71] | 2,478.9 ns |  5.30 ns |  4.43 ns |  0.90 | 0.1717 |      - |    2905 B |        0.92 |
| ParseFEN_Base2                        | rq2k2(...)- 0 1 [71] | 2,597.1 ns |  5.84 ns |  5.17 ns |  0.95 | 0.1793 |      - |    3025 B |        0.95 |
| ParseFEN_NoRegex                      | rq2k2(...)- 0 1 [71] | 1,271.5 ns |  9.36 ns |  8.76 ns |  0.46 | 0.0362 |      - |     624 B |        0.20 |
| ParseFEN_OptimizedParseBoard          | rq2k2(...)- 0 1 [71] |   397.2 ns |  1.51 ns |  1.26 ns |  0.14 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | rq2k2(...)- 0 1 [71] |   411.8 ns |  0.47 ns |  0.42 ns |  0.15 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | rq2k2(...)- 0 1 [71] |   346.1 ns |  3.78 ns |  3.54 ns |  0.13 | 0.0348 | 0.0010 |     584 B |        0.18 |


BenchmarkDotNet v0.14.0, Windows 10 (10.0.20348.4052) (Hyper-V)
AMD EPYC 7763, 1 CPU, 4 logical and 2 physical cores
.NET SDK 9.0.305
  [Host]     : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2

| Method                                | fen                  | Mean       | Error    | StdDev   | Ratio | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------------- |--------------------- |-----------:|---------:|---------:|------:|-------:|-------:|----------:|------------:|
| ParseFEN_Original                     | 8/k7/(...)- 0 1 [39] | 1,943.1 ns | 12.57 ns | 11.75 ns |  1.00 | 0.1755 |      - |    2961 B |        1.00 |
| ParseFEN_Improved1                    | 8/k7/(...)- 0 1 [39] | 1,856.3 ns | 14.68 ns | 13.73 ns |  0.96 | 0.1602 |      - |    2705 B |        0.91 |
| ParseFEN_Base2                        | 8/k7/(...)- 0 1 [39] | 1,880.6 ns | 11.08 ns |  9.82 ns |  0.97 | 0.1640 |      - |    2761 B |        0.93 |
| ParseFEN_NoRegex                      | 8/k7/(...)- 0 1 [39] |   806.5 ns |  2.16 ns |  2.02 ns |  0.42 | 0.0286 |      - |     480 B |        0.16 |
| ParseFEN_OptimizedParseBoard          | 8/k7/(...)- 0 1 [39] |   348.3 ns |  1.13 ns |  1.06 ns |  0.18 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedPopulateOccupancies | 8/k7/(...)- 0 1 [39] |   340.9 ns |  0.72 ns |  0.64 ns |  0.18 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedParseEnPassant      | 8/k7/(...)- 0 1 [39] |   280.1 ns |  2.10 ns |  1.96 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.20 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r2q1r(...)- 0 9 [68] | 2,447.4 ns | 14.02 ns | 12.43 ns |  1.00 | 0.1869 |      - |    3161 B |        1.00 |
| ParseFEN_Improved1                    | r2q1r(...)- 0 9 [68] | 2,334.1 ns | 10.08 ns |  9.43 ns |  0.95 | 0.1717 |      - |    2905 B |        0.92 |
| ParseFEN_Base2                        | r2q1r(...)- 0 9 [68] | 2,321.8 ns | 12.74 ns | 11.92 ns |  0.95 | 0.1793 |      - |    3017 B |        0.95 |
| ParseFEN_NoRegex                      | r2q1r(...)- 0 9 [68] | 1,084.8 ns |  2.54 ns |  2.25 ns |  0.44 | 0.0362 |      - |     624 B |        0.20 |
| ParseFEN_OptimizedParseBoard          | r2q1r(...)- 0 9 [68] |   422.6 ns |  1.00 ns |  0.88 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r2q1r(...)- 0 9 [68] |   407.0 ns |  1.12 ns |  1.04 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r2q1r(...)- 0 9 [68] |   330.7 ns |  2.53 ns |  2.36 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.18 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r3k2r(...)- 0 1 [68] | 2,343.7 ns | 11.56 ns | 10.82 ns |  1.00 | 0.1831 |      - |    3081 B |        1.00 |
| ParseFEN_Improved1                    | r3k2r(...)- 0 1 [68] | 2,275.1 ns | 13.01 ns | 12.17 ns |  0.97 | 0.1678 |      - |    2817 B |        0.91 |
| ParseFEN_Base2                        | r3k2r(...)- 0 1 [68] | 2,279.9 ns |  5.66 ns |  5.02 ns |  0.97 | 0.1717 |      - |    2929 B |        0.95 |
| ParseFEN_NoRegex                      | r3k2r(...)- 0 1 [68] | 1,004.1 ns |  1.78 ns |  1.66 ns |  0.43 | 0.0324 |      - |     552 B |        0.18 |
| ParseFEN_OptimizedParseBoard          | r3k2r(...)- 0 1 [68] |   387.9 ns |  1.20 ns |  1.06 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r3k2r(...)- 0 1 [68] |   372.9 ns |  1.37 ns |  1.22 ns |  0.16 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r3k2r(...)- 0 1 [68] |   320.6 ns |  2.62 ns |  2.46 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | r3k2r(...)- 0 1 [68] | 2,328.9 ns | 14.27 ns | 13.35 ns |  1.00 | 0.1831 |      - |    3081 B |        1.00 |
| ParseFEN_Improved1                    | r3k2r(...)- 0 1 [68] | 2,247.6 ns | 13.35 ns | 12.49 ns |  0.97 | 0.1678 |      - |    2817 B |        0.91 |
| ParseFEN_Base2                        | r3k2r(...)- 0 1 [68] | 2,221.3 ns | 10.52 ns |  9.84 ns |  0.95 | 0.1717 |      - |    2929 B |        0.95 |
| ParseFEN_NoRegex                      | r3k2r(...)- 0 1 [68] | 1,013.5 ns |  1.77 ns |  1.57 ns |  0.44 | 0.0324 |      - |     552 B |        0.18 |
| ParseFEN_OptimizedParseBoard          | r3k2r(...)- 0 1 [68] |   388.7 ns |  1.13 ns |  1.01 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | r3k2r(...)- 0 1 [68] |   389.4 ns |  0.69 ns |  0.61 ns |  0.17 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | r3k2r(...)- 0 1 [68] |   335.5 ns |  1.86 ns |  1.74 ns |  0.14 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rnbqk(...)6 0 1 [67] | 2,243.5 ns |  9.02 ns |  7.99 ns |  1.00 | 0.1831 |      - |    3073 B |        1.00 |
| ParseFEN_Improved1                    | rnbqk(...)6 0 1 [67] | 2,169.6 ns | 10.65 ns |  9.97 ns |  0.97 | 0.1640 |      - |    2801 B |        0.91 |
| ParseFEN_Base2                        | rnbqk(...)6 0 1 [67] | 2,187.3 ns |  9.47 ns |  8.86 ns |  0.97 | 0.1717 |      - |    2905 B |        0.95 |
| ParseFEN_NoRegex                      | rnbqk(...)6 0 1 [67] |   995.4 ns |  2.67 ns |  2.36 ns |  0.44 | 0.0305 |      - |     528 B |        0.17 |
| ParseFEN_OptimizedParseBoard          | rnbqk(...)6 0 1 [67] |   394.4 ns |  1.12 ns |  0.99 ns |  0.18 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | rnbqk(...)6 0 1 [67] |   397.1 ns |  1.12 ns |  0.99 ns |  0.18 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | rnbqk(...)6 0 1 [67] |   326.9 ns |  2.19 ns |  1.82 ns |  0.15 | 0.0348 | 0.0010 |     584 B |        0.19 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rnbqk(...)- 0 1 [56] | 1,720.7 ns |  9.79 ns |  8.68 ns |  1.00 | 0.1640 | 0.0019 |    2761 B |        1.00 |
| ParseFEN_Improved1                    | rnbqk(...)- 0 1 [56] | 1,601.6 ns |  9.60 ns |  8.98 ns |  0.93 | 0.1488 |      - |    2496 B |        0.90 |
| ParseFEN_Base2                        | rnbqk(...)- 0 1 [56] | 1,636.7 ns | 10.35 ns |  9.68 ns |  0.95 | 0.1545 |      - |    2584 B |        0.94 |
| ParseFEN_NoRegex                      | rnbqk(...)- 0 1 [56] |   633.6 ns |  1.22 ns |  1.08 ns |  0.37 | 0.0153 |      - |     264 B |        0.10 |
| ParseFEN_OptimizedParseBoard          | rnbqk(...)- 0 1 [56] |   361.1 ns |  0.79 ns |  0.74 ns |  0.21 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedPopulateOccupancies | rnbqk(...)- 0 1 [56] |   360.4 ns |  0.55 ns |  0.46 ns |  0.21 | 0.0100 |      - |     168 B |        0.06 |
| ParseFEN_OptimizedParseEnPassant      | rnbqk(...)- 0 1 [56] |   349.0 ns |  1.74 ns |  1.54 ns |  0.20 | 0.0348 | 0.0010 |     584 B |        0.21 |
|                                       |                      |            |          |          |       |        |        |           |             |
| ParseFEN_Original                     | rq2k2(...)- 0 1 [71] | 2,522.9 ns | 11.00 ns | 10.29 ns |  1.00 | 0.1869 |      - |    3169 B |        1.00 |
| ParseFEN_Improved1                    | rq2k2(...)- 0 1 [71] | 2,383.1 ns | 13.18 ns | 12.33 ns |  0.94 | 0.1717 |      - |    2905 B |        0.92 |
| ParseFEN_Base2                        | rq2k2(...)- 0 1 [71] | 2,419.7 ns | 13.33 ns | 12.47 ns |  0.96 | 0.1793 |      - |    3025 B |        0.95 |
| ParseFEN_NoRegex                      | rq2k2(...)- 0 1 [71] | 1,138.6 ns |  2.89 ns |  2.41 ns |  0.45 | 0.0362 |      - |     624 B |        0.20 |
| ParseFEN_OptimizedParseBoard          | rq2k2(...)- 0 1 [71] |   382.5 ns |  1.63 ns |  1.44 ns |  0.15 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedPopulateOccupancies | rq2k2(...)- 0 1 [71] |   371.8 ns |  1.22 ns |  1.14 ns |  0.15 | 0.0100 |      - |     168 B |        0.05 |
| ParseFEN_OptimizedParseEnPassant      | rq2k2(...)- 0 1 [71] |   328.4 ns |  2.23 ns |  2.09 ns |  0.13 | 0.0348 | 0.0010 |     584 B |        0.18 |


BenchmarkDotNet v0.14.0, macOS Ventura 13.7.6 (22H625) [Darwin 22.6.0]
Intel Core i7-8700B CPU 3.20GHz (Max: 3.19GHz) (Coffee Lake), 1 CPU, 4 logical and 4 physical cores
.NET SDK 9.0.305
  [Host]     : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2
  DefaultJob : .NET 9.0.9 (9.0.925.41916), X64 RyuJIT AVX2


| Method                                | fen                  | Mean       | Error     | StdDev    | Median     | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|-------------------------------------- |--------------------- |-----------:|----------:|----------:|-----------:|------:|--------:|-------:|-------:|----------:|------------:|
| ParseFEN_Original                     | 8/k7/(...)- 0 1 [39] | 2,777.7 ns |  45.24 ns |  40.11 ns | 2,757.0 ns |  1.00 |    0.02 | 0.4692 |      - |    2961 B |        1.00 |
| ParseFEN_Improved1                    | 8/k7/(...)- 0 1 [39] | 2,559.3 ns |  48.29 ns |  45.17 ns | 2,537.5 ns |  0.92 |    0.02 | 0.4311 |      - |    2705 B |        0.91 |
| ParseFEN_Base2                        | 8/k7/(...)- 0 1 [39] | 2,546.9 ns |  35.67 ns |  31.62 ns | 2,543.6 ns |  0.92 |    0.02 | 0.4387 |      - |    2761 B |        0.93 |
| ParseFEN_NoRegex                      | 8/k7/(...)- 0 1 [39] | 1,209.0 ns |   5.63 ns |   4.70 ns | 1,207.0 ns |  0.44 |    0.01 | 0.0763 |      - |     480 B |        0.16 |
…